@@ -1,87 +1,66 @@
 {
-  "name": "lions-adventskalender",
+  "name": "angular-nx",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
-    "nx": "nx",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "nx workspace-lint && ng lint",
-    "e2e": "ng e2e",
-    "affected:apps": "nx affected:apps",
-    "affected:libs": "nx affected:libs",
-    "affected:build": "nx affected:build",
-    "affected:e2e": "nx affected:e2e",
-    "affected:test": "nx affected:test",
-    "affected:lint": "nx affected:lint",
-    "affected:dep-graph": "nx affected:dep-graph",
-    "affected": "nx affected",
-    "format": "nx format:write",
-    "format:write": "nx format:write",
-    "format:check": "nx format:check",
-    "update": "nx migrate latest",
-    "workspace-generator": "nx workspace-generator",
-    "dep-graph": "nx dep-graph",
-    "help": "nx help"
+    "postinstall": "node ./decorate-angular-cli.js",
+    "start": "nx serve",
+    "build": "nx build",
+    "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.1.0",
-    "@angular/cdk": "^12.2.5",
-    "@angular/common": "^12.1.0",
-    "@angular/compiler": "^12.1.0",
-    "@angular/core": "^12.1.0",
-    "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "^12.1.0",
-    "@angular/material": "^12.2.5",
-    "@angular/platform-browser": "^12.1.0",
-    "@angular/platform-browser-dynamic": "^12.1.0",
-    "@angular/router": "^12.1.0",
-    "@angular/service-worker": "^12.1.0",
-    "@capacitor/core": "3.1.2",
-    "@ionic/angular": "^5.5.2",
+    "@angular/animations": "~14.2.0",
+    "@angular/cdk": "^14.2.5",
+    "@angular/common": "~14.2.0",
+    "@angular/compiler": "~14.2.0",
+    "@angular/core": "~14.2.0",
+    "@angular/flex-layout": "^14.0.0-beta.41",
+    "@angular/forms": "~14.2.0",
+    "@angular/material": "^14.2.5",
+    "@angular/platform-browser": "~14.2.0",
+    "@angular/platform-browser-dynamic": "~14.2.0",
+    "@angular/router": "~14.2.0",
+    "@angular/service-worker": "^14.2.7",
+    "@ionic/angular": "^6.3.3",
     "@ionic/storage-angular": "^3.0.6",
-    "@nrwl/angular": "12.6.4",
-    "cordova-sqlite-storage": "^6.0.0",
+    "@nrwl/angular": "15.0.3",
     "localforage-cordovasqlitedriver": "^1.8.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.1.0",
-    "@angular-eslint/eslint-plugin": "~12.3.0",
-    "@angular-eslint/eslint-plugin-template": "~12.3.0",
-    "@angular-eslint/template-parser": "~12.3.0",
-    "@angular/cli": "^12.0.0",
-    "@angular/compiler-cli": "^12.1.0",
-    "@angular/language-service": "^12.1.0",
-    "@ionic/angular-toolkit": "^4.0.0",
-    "@nrwl/cli": "12.6.4",
-    "@nrwl/cypress": "12.6.4",
-    "@nrwl/eslint-plugin-nx": "12.6.4",
-    "@nrwl/jest": "12.6.4",
-    "@nrwl/linter": "12.6.4",
+    "@angular-devkit/build-angular": "~14.2.0",
+    "@angular-eslint/eslint-plugin": "~14.0.4",
+    "@angular-eslint/eslint-plugin-template": "~14.0.4",
+    "@angular-eslint/template-parser": "~14.0.4",
+    "@angular/cli": "~14.2.0",
+    "@angular/compiler-cli": "~14.2.0",
+    "@angular/language-service": "~14.2.0",
+    "@nrwl/cli": "15.0.3",
+    "@nrwl/cypress": "15.0.3",
+    "@nrwl/eslint-plugin-nx": "15.0.3",
+    "@nrwl/jest": "15.0.3",
+    "@nrwl/linter": "15.0.3",
     "@nrwl/nx-cloud": "latest",
-    "@nrwl/tao": "12.6.4",
-    "@nrwl/workspace": "12.6.4",
-    "@types/jest": "26.0.24",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "~4.28.3",
-    "@typescript-eslint/parser": "~4.28.3",
-    "cypress": "^7.3.0",
-    "dotenv": "~10.0.0",
-    "eslint": "7.22.0",
+    "@nrwl/workspace": "15.0.3",
+    "@types/jest": "28.1.1",
+    "@types/node": "16.11.7",
+    "@typescript-eslint/eslint-plugin": "^5.36.1",
+    "@typescript-eslint/parser": "^5.36.1",
+    "cypress": "^10.7.0",
+    "eslint": "~8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.0.3",
-    "jest-preset-angular": "9.0.4",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.3",
-    "ts-node": "~9.1.1",
-    "typescript": "~4.3.5"
+    "jest": "28.1.1",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "~12.2.2",
+    "nx": "15.0.3",
+    "prettier": "^2.6.2",
+    "ts-jest": "28.0.5",
+    "ts-node": "10.9.1",
+    "typescript": "~4.8.2"
   }
 }
