@@ -8,6 +8,6 @@ describe('adventskalender', () => {
     cy.login('my-email@something.com', 'myPassword');
 
     // Function helper example, see `../support/app.po.ts` file
-    getGreeting().contains('Welcome to adventskalender!');
+    getGreeting().contains('Welcome adventskalender');
   });
 });
